fix: address code-review follow-ups (u16 enums, panic, typo, event names, perf)
- EnumFromPrimitive now uses u16 instead of u8 so integer enums whose
  values exceed 255 map to Unknown(..) instead of being unrepresentable
  (From<u16>/as_u16/serialize_u16); all deriving enums use Unknown(u16).
- OptionCreatedAt no longer panics on a malformed snowflake: replace
  .single().unwrap() inside map with and_then(.., .single()), matching
  the non-Option CreatedAt behaviour (returns None).
- Fix public variant typo ChannelStatues -> ChannelStatuses.
- Non-dispatch events now report a real event name from event_name()/
  Display (RECONNECT / HEARTBEAT_ACK / INVALID_SESSION) instead of the
  Rust variant identifier, matching dispatch events.
- Dispatch parsing deserializes directly from the payload Value instead
  of round-tripping through a String, saving one allocation per event.

+++ discord_client_structs/src/structs/application/mod.rs
@@ -181,7 +181,7 @@ pub enum ApplicationCommandType {
     ChatInput = 1,
     User = 2,
     Message = 3,
-    Unknown(u8),
+    Unknown(u16),
 }
 
 #[derive(Debug, Clone, PartialEq, Eq, EnumFromPrimitive)]
@@ -192,7 +192,7 @@ pub enum ApplicationType {
     Music = 2, // Should be disabled by discord
     TicketedEvents = 3,
     CreatorMonetization = 4,
-    Unknown(u8),
+    Unknown(u16),
 }
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, EnumFromString)]
